Pulled createIconArray into its own folder in utils
