+++ src/components/StarRating/index.js
@@ -1,8 +1,12 @@
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
+import { createIconArray } from "../../utils/createIconArray";
+
 const StarRating = (props) => {
-  const { iconArray } = props;
+  const { rating } = props;
+  const iconArray = createIconArray(rating);
+
   return (
     <div>
       {iconArray.map((icon, index) => {
