+++ src/containers/RestaurantListContainer/restaurantListContainerUtils.js
@@ -1,5 +1,3 @@
-import { faStar, faStarHalf } from "@fortawesome/free-solid-svg-icons";
-
 import { LabelsObject } from "../../Labels";
 
 const checkValidImageURL = (url) => {
@@ -12,19 +10,8 @@ const checkValidImageURL = (url) => {
   }
 };
 
-const createIconArray = (rating) => {
-  const iconArray = [];
-  for (let i = 0; i < Math.floor(rating); i++) {
-    iconArray.push(faStar);
-  }
-  if (rating % 1 !== 0) {
-    iconArray.push(faStarHalf);
-  }
-  return iconArray;
-};
-
 const formattedAddress = (address) => {
   return address.replace(/\n/g, ", ");
 };
 
-export { checkValidImageURL, createIconArray, formattedAddress };
+export { checkValidImageURL, formattedAddress };
