+++ src/containers/RestaurantListContainer/restaurantListContainerUtils.test.js
@@ -1,8 +1,5 @@
-import { faStar, faStarHalf } from "@fortawesome/free-solid-svg-icons";
-
 import {
   checkValidImageURL,
-  createIconArray,
   formattedAddress,
 } from "./restaurantListContainerUtils";
 import { LabelsObject } from "../../Labels";
@@ -17,22 +14,6 @@ describe("checkValidImageURL tests", () => {
     const { invalidImageURLLabel } = LabelsObject;
     const invalidURL = "https://s3-media3.fl.yelpcdn.com/bphoto/None/o.jpg";
     expect(checkValidImageURL(invalidURL)).toEqual(invalidImageURLLabel);
-  });
-});
-
-describe("checkCorrectStarRating", () => {
-  test("Returns star rating array for whole number", () => {
-    const wholeNumberArray = [faStar, faStar];
-    expect(createIconArray(2)).toEqual(
-      expect.arrayContaining(wholeNumberArray)
-    );
-  });
-
-  test("Returns star rating array for half numbers", () => {
-    const halfNumberArray = [faStar, faStarHalf];
-    expect(createIconArray(1.5)).toEqual(
-      expect.arrayContaining(halfNumberArray)
-    );
   });
 });
